requirements.txt
Django >=3.2.4,<3.3
djangorestframework >=3.12.4,<3.13
psycopg2 >=2.9.1,<2.10
drf-spectacular >=0.21.1,<0.22